+++ packages/pluggableWidgets/line-chart-native/typings/LineChartProps.d.ts
@@ -35,8 +35,8 @@ export interface LinesType {
 
 export interface LinesPreviewType {
     dataSet: DataSetEnum;
-    staticDataSource: {} | { type: string } | null;
-    dynamicDataSource: {} | { type: string } | null;
+    staticDataSource: {} | { caption: string } | { type: string } | null;
+    dynamicDataSource: {} | { caption: string } | { type: string } | null;
     groupByAttribute: string;
     staticName: string;
     dynamicName: string;