+++ packages/pluggableWidgets/pie-doughnut-chart-native/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [2.0.1] - 2023-5-17
+
+### Fixed
+
+-   We fixed an issue where height or width with a value of 0 crash the app.
+
 ## [2.0.0] - 2023-3-28
 
 ## BREAKING